+++ media-attached-filter.php
@@ -114,9 +114,17 @@ function media_attached_filter_search_ajax(): void {
 		wp_send_json( array( 'success' => false ) );
 	}
 
+	// get post types.
+	$post_types = media_attached_filter_get_post_types();
+
+	// bail if list is empty.
+	if ( empty( $post_types ) ) {
+		wp_send_json( array( 'success' => false ) );
+	}
+
 	// define query.
 	$query  = array(
-		'post_type'      => array( 'post', 'page' ),
+		'post_type'      => $post_types,
 		'post_status'    => 'any',
 		's'              => $keyword,
 		'fields'         => 'ids',
@@ -133,7 +141,7 @@ function media_attached_filter_search_ajax(): void {
 		}
 
 		// add the entry to the resulting list.
-		$list[ absint( $post_id ) ] = get_the_title( $post_id );
+		$list[ absint( $post_id ) ] = get_post_field( 'post_title', $post_id, 'raw' );
 	}
 
 	// return resulting list.
@@ -165,7 +173,7 @@ function media_attached_filter_run_filter( WP_Query $query ): void {
 	}
 
 	// get the attribute from request.
-	$attached = filter_input( INPUT_GET, 'maf_attached', FILTER_SANITIZE_FULL_SPECIAL_CHARS );
+	$attached = filter_input( INPUT_GET, 'maf_attached', FILTER_SANITIZE_SPECIAL_CHARS );
 
 	// bail if attribute is not set.
 	if ( is_null( $attached ) ) {
@@ -177,9 +185,17 @@ function media_attached_filter_run_filter( WP_Query $query ): void {
 		return;
 	}
 
+	// get post types.
+	$post_types = media_attached_filter_get_post_types();
+
+	// bail if list is empty.
+	if ( empty( $post_types ) ) {
+		return;
+	}
+
 	// query for the attached page or post.
 	$query_to_get_post_id = array(
-		'post_type'   => array( 'post', 'page' ),
+		'post_type'   => $post_types,
 		'post_status' => 'any',
 		'title'       => $attached,
 		'fields'      => 'ids',
@@ -195,3 +211,45 @@ function media_attached_filter_run_filter( WP_Query $query ): void {
 	}
 }
 add_action( 'pre_get_posts', 'media_attached_filter_run_filter' );
+
+/**
+ * Add links in row meta.
+ *
+ * @param array<string,string> $links List of links.
+ * @param string               $file The requested plugin file name.
+ *
+ * @return array<string,string>
+ */
+function media_attached_filter_add_row_meta_links( array $links, string $file ): array {
+	// bail if this is not our plugin.
+	if ( __FILE__ !== WP_PLUGIN_DIR . DIRECTORY_SEPARATOR . $file ) {
+		return $links;
+	}
+
+	// add our custom links.
+	$row_meta = array(
+		'support' => '<a href="https://wordpress.org/support/plugin/media-attached-filter/" target="_blank" title="' . esc_html__( 'Support Forum', 'media-attached-filter' ) . '">' . esc_html__( 'Support Forum', 'media-attached-filter' ) . '</a>',
+	);
+
+	// return the resulting list of links.
+	return array_merge( $links, $row_meta );
+}
+add_filter( 'plugin_row_meta', 'media_attached_filter_add_row_meta_links', 10, 2 );
+
+/**
+ * Return list of post types we support.
+ *
+ * @return array<int,string>
+ */
+function media_attached_filter_get_post_types(): array {
+	// get post types.
+	$post_types = get_post_types( array( 'public' => true ) );
+
+	/**
+	 * Filter the supported post types.
+	 *
+	 * @since 1.2.1 Available since 1.2.1.
+	 * @param array<int,string> $post_types List of post types.
+	 */
+	return apply_filters( 'media_attached_filter_post_types', $post_types );
+}