+++ changelog.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [1.2.1] - 04.10.2025
+
+### Added
+
+- Added hook for supported post types
+
+### Changed
+
+- Use only public post types
+
+## [1.2.0] - 03.10.2025
+
+### Added
+
+- Added support for every post type
+- Added link to support forum in plugin list
+
+### Changed
+
+- Optimized encoding of characters in titles in request and response
+
 ## [1.1.0] - 15.05.2025
 
 ### Added